[Meeting] Modify the id from an int to a long

+++ app/src/main/java/com/example/mareu/model/Meeting.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 public class Meeting {
 
 
-    private int id;
+    private long id;
     private String name;
     private LocalTime startTime;
     private LocalTime endTime;
@@ -18,7 +18,7 @@ public class Meeting {
     private List<String> participants;
 
 
-    public Meeting(int id, String name, LocalTime startTime, LocalTime endTime, LocalDate date, String location, String subject, List<String> participants) {
+    public Meeting(long id, String name, LocalTime startTime, LocalTime endTime, LocalDate date, String location, String subject, List<String> participants) {
         this.id = id;
         this.name = name;
         this.startTime = startTime;
@@ -31,11 +31,11 @@ public class Meeting {
 
     // --- GETTERS AND SETTERS ---
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
